+++ src/dictionary/DateEdit.ts
@@ -5,7 +5,7 @@ export interface DateEditInitOption extends SimpleDateEditInitOption {
   type: 'date'
 }
 
-class DateEdit extends SimpleDateEdit{
+class DateEdit extends SimpleDateEdit<false>{
   static $name = 'DateEdit'
   type: 'date'
   constructor(initOption: DateEditInitOption, parent?: DictionaryValue, modName?: string) {